requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.10.0
